requirements.txt
requests
urllib3
dnacentersdk==2.0.2
flask
Flask-BasicAuth
slackclient==2.9.3